feat: clean ups
